Organize ui components files under view_component directory.
