+++ app/helpers/view_component/course_card_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module ViewComponent
+  module CourseCardComponent
+    def course_card_component(course:, enrollment: nil)
+      render partial: 'view_components/app/course_carousal/course_card_component', locals: { course:, enrollment: }
+    end
+  end
+end
