+++ app/helpers/view_component/icon_component.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module ViewComponent
+  module IconComponent
+    def icon(icon_name, css: nil, span_css: nil)
+      file = Rails.cache.fetch("icon:#{icon_name}") do
+        Rails.root.join('app', 'assets', 'icons', "#{icon_name}.svg").read
+      end
+
+      svg = Nokogiri::HTML::DocumentFragment.parse(file).at_css('svg')
+      svg['class'] = css
+      content_tag(:span, svg.to_html.html_safe, class: "inline-flex justify-center items-center #{span_css}".strip) # rubocop:disable Rails/OutputSafety
+    end
+  end
+end
