+++ app/views/learning_partners/index.html.erb
@@ -21,5 +21,5 @@
     <% end %>
   </ul>
   <div class="mx-4 my-6 text-center">
-    <%= paginator_component(@learning_partners, path: learning_partners_path) %>
+    <%= paginator_component(collection: @learning_partners, path: learning_partners_path) %>
   </div>
