--- app/views/ui/paginator/_paginator_component.html.erb
@@ -1,6 +1,0 @@
-<% if collection.prev_page || collection.next_page %>
-  <div class="flex justify-center mt-4 space-x-2">
-    <%= render "ui/paginator/prev_page", collection: collection, path: path %>
-    <%= render "ui/paginator/next_page", collection: collection, path: path %>
-  </div>
-<% end %>
