+++ app/views/view_components/app/course_carousal/_course_card_component.html.erb
