+++ app/views/view_components/app/course_carousal/_course_carousal_component.html.erb
