+++ app/views/view_components/app/course_select/_search_component.html.erb
