+++ app/views/view_components/app/course_select/_sidebar_component.html.erb
