+++ app/views/view_components/icons/index.html.erb
