+++ app/views/view_components/modals/_modal_component.html.erb
