+++ app/views/view_components/notification_bars/_notification_bar.html.erb
