+++ app/views/view_components/notification_bars/index.html.erb
