+++ app/views/view_components/paginator/_next_page.html.erb
