+++ app/views/view_components/tables/index.html.erb
