+++ app/views/view_components/typography/_heading_component.html.erb
