+++ app/views/view_components/typography/_linked_text_component.html.erb
