+++ config/locales/ui.en.yml
@@ -1,4 +1,4 @@
 en:
   components:
     title: "Components"
-    label: "View ui components"
+    label: "View components"
